services/raw_ingest/requirements.txt
aiokafka==0.9.0
kafka-python==2.0.2
asyncpg==0.29.0
psycopg2-binary==2.9.9
